+++ src/parser/inline.ts
@@ -69,5 +69,5 @@ export { indexer, index } from './inline/extension/indexer';
 export { label, isGroup, isFixed } from './inline/extension/label';
 export { link } from './inline/link';
 export { media } from './inline/media';
-export { uri } from './inline/autolink/uri';
+export { uri, address, attribute } from './inline/autolink/uri';
 export { shortmedia } from './inline/shortmedia';